+++ site_scons/site_tools/NVDATool/typings.py
@@ -20,19 +20,25 @@ class AddonInfo(TypedDict):
 
 
 class BrailleTableAttributes(TypedDict):
-    displayName: str
-    contracted: bool
-    output: bool
-    input: bool
+	displayName: str
+	contracted: bool
+	output: bool
+	input: bool
 
 
 class SymbolDictionaryAttributes(TypedDict):
-    displayName: str
-    mandatory: bool
+	displayName: str
+	mandatory: bool
+
+
+class SpeechDictionaryAttributes(TypedDict):
+	displayName: str
+	mandatory: bool
 
 
 BrailleTables = dict[str, BrailleTableAttributes]
 SymbolDictionaries = dict[str, SymbolDictionaryAttributes]
+SpeechDictionaries = dict[str, SpeechDictionaryAttributes]
 
 
 class Strable(Protocol):