+++ addon/appModules/nvcplui.py
@@ -3,8 +3,7 @@
 # See the file COPYING.txt for more details.
 # Copyright (C) 2024 hwf1324 <1398969445@qq.com>
 
-"""NVIDIA Control Panel appModule.
-"""
+"""NVIDIA Control Panel appModule."""
 
 import appModuleHandler
 import controlTypes
@@ -12,7 +11,6 @@ from NVDAObjects import NVDAObject
 
 
 class AppModule(appModuleHandler.AppModule):
-
 	def event_NVDAObject_init(self, obj: NVDAObject):
 		if controlTypes.Role.STATICTEXT == obj.role:
 			obj.name = obj.windowText