+++ changelog.md
@@ -3,3 +3,4 @@
 * Add mouse tracking support for frameless windows in the Tauri framework.
   * Thanks to @ryusei-48 for providing the patch reference.
 * Disabled the forced use of UIA implementation in the 飞书 module and fixed the character navigation in the edit box. Supported mouse tracking using redirection objects.
+* Update add-on template.