+++ changelog.md
@@ -1,3 +1,6 @@
-# v0.9.1
+# v0.9.2
 
-* Use UIAHandler's MTA thread for UIA queries.
+* Add mouse tracking support for frameless windows in the Tauri framework.
+  * Thanks to @ryusei-48 for providing the patch reference.
+* Disabled the forced use of UIA implementation in the 飞书 module and fixed the character navigation in the edit box. Supported mouse tracking using redirection objects.
+* Update add-on template.